pkg/otlp: move to comp/otelcol/otlp
This change moves the pkg/otlp package to comp/otelcol/otlp as it
represents the core code for this component. Some other package still
rely on it (pkg/status, pkg/serverless and pkg/metadata/host) but this
dependency should be removable once those pacakges are migrated to
components.
