+++ comp/otelcol/otlp/collector_test_config.go
@@ -3,11 +3,13 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2021-present Datadog, Inc.
 
-//go:build otlp && serverless && test
+//go:build otlp && !serverless && test
 
 package otlp
 
-import "github.com/DataDog/datadog-agent/pkg/otlp/internal/testutil"
+import (
+	"github.com/DataDog/datadog-agent/comp/otelcol/otlp/internal/testutil"
+)
 
 func getTestPipelineConfig() PipelineConfig {
 	return PipelineConfig{
@@ -15,7 +17,7 @@ func getTestPipelineConfig() PipelineConfig {
 		TracePort:          5003,
 		MetricsEnabled:     true,
 		TracesEnabled:      true,
-		LogsEnabled:        false,
+		LogsEnabled:        true,
 		Metrics:            map[string]interface{}{},
 	}
 }
