+++ comp/otelcol/otlp/collector_test_config_serverless.go
@@ -3,13 +3,11 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2021-present Datadog, Inc.
 
-//go:build otlp && !serverless && test
+//go:build otlp && serverless && test
 
 package otlp
 
-import (
-	"github.com/DataDog/datadog-agent/pkg/otlp/internal/testutil"
-)
+import "github.com/DataDog/datadog-agent/comp/otelcol/otlp/internal/testutil"
 
 func getTestPipelineConfig() PipelineConfig {
 	return PipelineConfig{
@@ -17,7 +15,7 @@ func getTestPipelineConfig() PipelineConfig {
 		TracePort:          5003,
 		MetricsEnabled:     true,
 		TracesEnabled:      true,
-		LogsEnabled:        true,
+		LogsEnabled:        false,
 		Metrics:            map[string]interface{}{},
 	}
 }
