+++ comp/otelcol/otlp/example/metric/README.md
