+++ comp/otelcol/otlp/example/metric/doc.go
