+++ comp/otelcol/otlp/example/metric/go.mod
