+++ comp/otelcol/otlp/example/metric/main.go
