+++ comp/otelcol/otlp/internal/configutils/testdata/pipeline.yaml
