+++ comp/otelcol/otlp/internal/configutils/utils.go
