+++ comp/otelcol/otlp/internal/serializerexporter/config.go
