+++ comp/otelcol/otlp/internal/testutil/doc.go
