+++ comp/otelcol/otlp/internal/testutil/testutil.go
