+++ comp/otelcol/otlp/map_provider_not_serverless_test.go
@@ -18,7 +18,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/collector/confmap"
 
-	"github.com/DataDog/datadog-agent/pkg/otlp/internal/testutil"
+	"github.com/DataDog/datadog-agent/comp/otelcol/otlp/internal/testutil"
 	"github.com/DataDog/datadog-agent/pkg/serializer"
 )
