+++ comp/otelcol/otlp/testdata/logs_enabled.yaml
