+++ comp/otelcol/otlp/testdata/metrics/allconfig.yaml
