+++ comp/otelcol/otlp/testdata/receiver/advanced.yaml
