+++ comp/otelcol/otlp/testdata/receiver/noprotocols.yaml
