+++ comp/otelcol/otlp/testdata/receiver/null.yaml
