+++ comp/otelcol/otlp/testdata/receiver/simple.yaml
