+++ comp/otelcol/otlp/testdata/receiver/typo.yaml
