--- pkg/otlp/testdata/debug/empty_but_set_debug.yaml
@@ -1,2 +1,0 @@
-otlp_config:
-  debug:
