--- pkg/otlp/testdata/debug/loglevel_debug.yaml
@@ -1,3 +1,0 @@
-otlp_config:
-  debug:
-    loglevel: debug
