--- pkg/otlp/testdata/debug/loglevel_disabled.yaml
@@ -1,3 +1,0 @@
-otlp_config:
-  debug:
-    loglevel: disabled
