--- pkg/otlp/testdata/debug/verbosity_normal.yaml
@@ -1,3 +1,0 @@
-otlp_config:
-  debug:
-    verbosity: normal
