+++ pkg/serverless/otlp/otlp.go
@@ -14,8 +14,8 @@ import (
 
 	"go.opentelemetry.io/collector/otelcol"
 
+	coreOtlp "github.com/DataDog/datadog-agent/comp/otelcol/otlp"
 	"github.com/DataDog/datadog-agent/pkg/config"
-	coreOtlp "github.com/DataDog/datadog-agent/pkg/otlp"
 	"github.com/DataDog/datadog-agent/pkg/serializer"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
